+++ demos/randomPolygon.js
@@ -3,11 +3,11 @@
 // |
 // | @date     2026-09-15
 // +-------------------------------
-var randomPolygon = function (viewport, pointCount, isClockwise) {
+var randomPolygon = function (viewport, pointCount, isCounterClockwise) {
   var radius = viewport.getMinDimension() / 2.0;
   var center = viewport.getCenter();
   var points = [];
-  var directionFactor = isClockwise ? -1 : 1;
+  var directionFactor = isCounterClockwise ? -1 : 1;
   for (var i = 0; i < pointCount; i++) {
     var point = new Vertex(radius * 0.1 + Math.random() * 0.9 * radius, 0.0);
     point.rotate(directionFactor * (Math.PI / pointCount) * 2 * i, center);