+++ src/cjs/VertTuple.d.ts
@@ -17,13 +17,15 @@
  * @modified 2025-04-15 Changed param of `VertTuple.moveTo` method from `Vertex` to `XYCoords`.
  * @modified 2025-04-15 Added method `VertTuple.move` method.
  * @modified 2026-06-10 Adding helper function `VertTuple.utils.calcCircumcircle`.
- * @modified 2026-06-17 Adding method `VertTuple.asLine` for converting Vector to Line instances.
- * @version 1.6.0
+ * @modified 2026-06-17 Adding method `VertTuple.asLine` for converting TertTuples to Line instances.
+ * @modified 2026-09-21 Adding method `VertTuple.asVector` for converting TertTuples to Vector instances.
+ * @version 1.7.0
  */
 import { Vertex } from "./Vertex";
 import { XYCoords, UID } from "./interfaces";
 import { ICircle } from "./Circle";
 import { Line } from "./Line";
+import { Vector } from "./Vector";
 /**
  * @classdesc An abstract base classes for vertex tuple constructs, like Lines or Vectors.
  * @abstract
@@ -265,14 +267,23 @@ export declare class VertTuple<T extends VertTuple<T>> {
      **/
     clone(): T;
     /**
-     * Converts this `Vector` to a `Line` (segment).
+     * Converts this `VertTuple` to a `Line` (segment).
      *
      * @method asLine
-     * @return {T} A type safe clone if this instance.
+     * @return {Line} This tuple as `Line` instance.
      * @instance
      * @memberof VertTuple
      **/
     asLine(): Line;
+    /**
+     * Converts this `VertTuple` to a `Vector` (arrow).
+     *
+     * @method asVector
+     * @return {Line} This tuple as `Vector` instance.
+     * @instance
+     * @memberof VertTuple
+     **/
+    asVector(): Vector;
     /**
      * Create a string representation of this line.
      *