+++ src/esm/Polygon.d.ts
@@ -39,6 +39,7 @@
  * @modified 2025-05-20 Tweaking `Polygon.getInnerAngleAt` and `Polygo.isAngleAcute` to handle indices out of array bounds as well.
  * @modified 2025-06-07 Adding `Polygon.closestLineIntersectionIndex` to determine line intersections plus detected edge index.
  * @modified 2026-09-16 Adding a `forceClockwise` parameter to the `Polygon.getCentroid()` method.
+ * @modified 2026-09-21 Adding `Polygon.revert` method to change the winding order.
  * @version 1.17.0
  *
  * @file Polygon
@@ -262,6 +263,16 @@ export declare class Polygon implements IBounded, Intersectable, SVGSerializable
      * @return {boolean}
      */
     isClockwise(): boolean;
+    /**
+     * Revert the order of this polygon's vertices to change the winding order.
+     * This operation is in-place.
+     *
+     * @method revert
+     * @instance
+     * @memberof Polygon
+     * @return {Polygon} This for chaining.
+     */
+    revert(): Polygon;
     /**
      * Get the perimeter of this polygon.
      * The perimeter is the absolute length of the outline.