+++ src/ts/VertTuple.ts
@@ -17,8 +17,9 @@
  * @modified 2025-04-15 Changed param of `VertTuple.moveTo` method from `Vertex` to `XYCoords`.
  * @modified 2025-04-15 Added method `VertTuple.move` method.
  * @modified 2026-06-10 Adding helper function `VertTuple.utils.calcCircumcircle`.
- * @modified 2026-06-17 Adding method `VertTuple.asLine` for converting Vector to Line instances.
- * @version 1.6.0
+ * @modified 2026-06-17 Adding method `VertTuple.asLine` for converting TertTuples to Line instances.
+ * @modified 2026-09-21 Adding method `VertTuple.asVector` for converting TertTuples to Vector instances.
+ * @version 1.7.0
  */
 
 import { Vertex } from "./Vertex";
@@ -26,6 +27,7 @@ import { UIDGenerator } from "./UIDGenerator";
 import { XYCoords, UID } from "./interfaces";
 import { ICircle } from "./Circle";
 import { Line } from "./Line";
+import { Vector } from "./Vector";
 
 /**
  * @classdesc An abstract base classes for vertex tuple constructs, like Lines or Vectors.
@@ -372,15 +374,27 @@ export class VertTuple<T extends VertTuple<T>> {
   }
 
   /**
-   * Converts this `Vector` to a `Line` (segment).
+   * Converts this `VertTuple` to a `Line` (segment).
    *
    * @method asLine
-   * @return {T} A type safe clone if this instance.
+   * @return {Line} This tuple as `Line` instance.
    * @instance
    * @memberof VertTuple
    **/
   asLine(): Line {
     return new Line(this.a, this.b);
+  }
+
+  /**
+   * Converts this `VertTuple` to a `Vector` (arrow).
+   *
+   * @method asVector
+   * @return {Line} This tuple as `Vector` instance.
+   * @instance
+   * @memberof VertTuple
+   **/
+  asVector(): Vector {
+    return new Vector(this.a, this.b);
   }
 
   /**