+++ src/ts/utils/renderers/VoronoiRenderer.ts
@@ -10,6 +10,7 @@
 import { Bounds } from "../../Bounds";
 import { Polygon } from "../../Polygon";
 import { Triangle } from "../../Triangle";
+import { DrawLib } from "../../interfaces";
 import { convexPolygonIncircle } from "../algorithms/convexPolygonIncircle";
 import { sutherlandHodgman } from "../algorithms/sutherlandHodgman";
 import { cloneVertexArray } from "../cloneVertexArray";
@@ -64,7 +65,7 @@ export class VoronoiRenderer {
    * @param {string} color
    * @return {void}
    */
-  draw(draw, fill) {
+  draw(draw: DrawLib<any>, fill: DrawLib<any>) {
     // Draw circumcircles
     if (this.voronoiContext.config.drawCircumCircles) {
       VoronoiRenderer.drawCircumCircles(draw, this.voronoiContext.triangles);
@@ -82,9 +83,21 @@ export class VoronoiRenderer {
     }
   }
 
-  private _drawCell(draw, fill, clipBoxPolygon: Polygon, cell) {
-    var polygon = cell.toPolygon();
-    polygon.scale(this.voronoiContext.config.voronoiCellScale, cell.sharedVertex);
+  /**
+   * Draw the given triangle with the specified (CSS-) color.
+   *
+   * @name _drawCell
+   * @instance
+   * @private
+   * @memberof VoronoiRenderer
+   * @param {DrawLib} draw
+   * @param {Polygon} clipBoxPolygon
+   * @param {VoronoiCell} cell
+   * @return {void}
+   */
+  private _drawCell(draw: DrawLib<any>, fill: DrawLib<any>, clipBoxPolygon: Polygon, cell: VoronoiCell) {
+    var cellPolygon = cell.toPolygon();
+    cellPolygon.scale(this.voronoiContext.config.voronoiCellScale, cell.sharedVertex);
 
     // Draw large (unclipped) Voronoi cell
     if (
@@ -92,7 +105,7 @@ export class VoronoiRenderer {
       (!this.voronoiContext.config.clipVoronoiCells || this.voronoiContext.config.drawUnclippedVoronoiCells)
     ) {
       draw.polyline(
-        polygon.vertices,
+        cellPolygon.vertices,
         false,
         this.voronoiContext.config.clipVoronoiCells ? "rgba(128,128,128,0.333)" : this.voronoiContext.config.voronoiOutlineColor,
         this.voronoiContext.config.voronoiCellLineWidth ?? 2.0
@@ -102,12 +115,13 @@ export class VoronoiRenderer {
     // Apply clipping?
     if (this.voronoiContext.config.clipVoronoiCells) {
       // Clone the array here: convert Array<XYCoords> to Array<Vertex>
-      polygon = new Polygon(cloneVertexArray(sutherlandHodgman(polygon.vertices, clipBoxPolygon.vertices)), false);
+      // polygon = new Polygon(cloneVertexArray(sutherlandHodgman(polygon.vertices, clipBoxPolygon.vertices)), false);
+      cellPolygon = VoronoiRenderer.clipVoronoiPolygon(cellPolygon, clipBoxPolygon);
     }
 
     if (this.voronoiContext.config.drawVoronoiOutlines && this.voronoiContext.config.clipVoronoiCells) {
       draw.polygon(
-        polygon,
+        cellPolygon,
         this.voronoiContext.config.voronoiOutlineColor,
         this.voronoiContext.config.voronoiCellLineWidth ?? 2.0
       );
@@ -115,7 +129,7 @@ export class VoronoiRenderer {
 
     if ((!cell.isOpen() || this.voronoiContext.config.clipVoronoiCells) && cell.triangles.length >= 3) {
       if (this.voronoiContext.config.drawCubicCurves) {
-        var cbezier = polygon.toCubicBezierData(this.voronoiContext.config.voronoiCubicThreshold);
+        var cbezier = cellPolygon.toCubicBezierData(this.voronoiContext.config.voronoiCubicThreshold);
         if (this.voronoiContext.config.fillVoronoiCells) {
           fill.cubicBezierPath(cbezier, this.voronoiContext.config.voronoiCellColor);
         } else {
@@ -123,9 +137,9 @@ export class VoronoiRenderer {
         }
       }
       if (this.voronoiContext.config.drawVoronoiIncircles) {
-        var result = convexPolygonIncircle(polygon);
+        var result = convexPolygonIncircle(cellPolygon);
         var circle = result.circle;
-        var triangle = result.triangle;
+        // var triangle = result.triangle;
         // Here we should have found the best inlying circle (and the corresponding triangle)
         // inside the Voronoi cell.
         draw.circle(circle.center, circle.radius, "rgba(255,192,0,1.0)", 2);
@@ -152,26 +166,35 @@ export class VoronoiRenderer {
    * @name drawTriangle
    * @memberof VoronoiRenderer
    * @param {DrawLib} draw
-   * @param {Triangle} t
+   * @param {Triangle} tri
    * @param {string} color
    * @return {void}
    */
-  static drawTriangle(draw, t, color) {
+  static drawTriangle(draw: DrawLib<any>, tri: Triangle, color: string) {
     // draw.line(t.a, t.b, color);
     // draw.line(t.b, t.c, color);
     // draw.line(t.c, t.a, color);
-    draw.polyline([t.a, t.b, t.c], false, color);
+    draw.polyline([tri.a, tri.b, tri.c], false, color);
   }
 
   /**
    * Draw the circumcircles of all triangles.
    */
-  static drawCircumCircles(draw, triangles) {
+  static drawCircumCircles(draw: DrawLib<any>, triangles: Array<Triangle>) {
     for (var t in triangles) {
       var cc = triangles[t].getCircumcircle();
       draw.circle(cc.center, cc.radius, "#e86800");
     }
   }
 
-  // _context.VoronoiRenderer = VoronoiRenderer;
+  /**
+   * Clip the convex (!) cell polygon by the convex (!) clipping polygon.
+   *
+   * @param cellPolygon
+   * @param clipBoxPolygon
+   * @returns
+   */
+  static clipVoronoiPolygon(cellPolygon: Polygon, clipBoxPolygon: Polygon) {
+    return new Polygon(cloneVertexArray(sutherlandHodgman(cellPolygon.vertices, clipBoxPolygon.vertices)), false);
+  }
 }